Improve pub/sub notification usage
