--- DiscordTranslationBot/Commands/SlashCommandExecuted/ProcessTranslateSlashCommand.cs
@@ -1,14 +1,0 @@
-﻿using Discord;
-
-namespace DiscordTranslationBot.Commands.SlashCommandExecuted;
-
-/// <summary>
-/// Command to process a translate slash command.
-/// </summary>
-public sealed class ProcessTranslateSlashCommand : IRequest
-{
-    /// <summary>
-    /// The slash command from the event.
-    /// </summary>
-    public required ISlashCommandInteraction Command { get; init; }
-}
