--- DiscordTranslationBot/Commands/SlashCommandExecuted/RegisterSlashCommands.cs
@@ -1,12 +1,0 @@
-﻿using Discord;
-
-namespace DiscordTranslationBot.Commands.SlashCommandExecuted;
-
-/// <summary>
-/// Command for registering slash commands.
-/// </summary>
-public sealed class RegisterSlashCommands : IRequest, IRegisterCommands
-{
-    /// <inheritdoc cref="IRegisterCommands.Guild" />
-    public IGuild? Guild { get; init; }
-}
