--- DiscordTranslationBot/Handlers/JoinedGuildHandler.cs
@@ -1,34 +1,0 @@
-﻿using DiscordTranslationBot.Commands.MessageCommandExecuted;
-using DiscordTranslationBot.Commands.SlashCommandExecuted;
-using DiscordTranslationBot.Notifications;
-
-namespace DiscordTranslationBot.Handlers;
-
-/// <summary>
-/// Handler for the Joined Guild Discord event.
-/// </summary>
-public sealed class JoinedGuildHandler : INotificationHandler<JoinedGuildNotification>
-{
-    private readonly IMediator _mediator;
-
-    /// <summary>
-    /// Instantiates a new instance of the <see cref="JoinedGuildHandler" /> class.
-    /// </summary>
-    /// <param name="mediator">Mediator to use.</param>
-    public JoinedGuildHandler(IMediator mediator)
-    {
-        _mediator = mediator;
-    }
-
-    /// <summary>
-    /// Delegates joined guild events to the correct handler.
-    /// </summary>
-    /// <param name="notification">The notification.</param>
-    /// <param name="cancellationToken">The cancellation token.</param>
-    public async Task Handle(JoinedGuildNotification notification, CancellationToken cancellationToken)
-    {
-        await _mediator.Send(new RegisterMessageCommands { Guild = notification.Guild }, cancellationToken);
-
-        await _mediator.Send(new RegisterSlashCommands { Guild = notification.Guild }, cancellationToken);
-    }
-}
