+++ DiscordTranslationBot/Handlers/TranslateMessageCommandHandler.cs
@@ -1,7 +1,4 @@
-﻿using System.Text.Json;
-using Discord;
-using Discord.Net;
-using DiscordTranslationBot.Commands.MessageCommandExecuted;
+﻿using Discord;
 using DiscordTranslationBot.Constants;
 using DiscordTranslationBot.Models.Providers.Translation;
 using DiscordTranslationBot.Notifications;
@@ -12,33 +9,26 @@ using Humanizer;
 namespace DiscordTranslationBot.Handlers;
 
 /// <summary>
-/// Handles the MessageCommandExecuted event of the Discord client.
+/// Handler for the translate message command.
 /// </summary>
-public partial class MessageCommandExecutedHandler
-    : INotificationHandler<MessageCommandExecutedNotification>,
-        IRequestHandler<RegisterMessageCommands>,
-        IRequestHandler<ProcessTranslateMessageCommand>
+public partial class TranslateMessageCommandHandler : INotificationHandler<MessageCommandExecutedNotification>
 {
     private readonly IDiscordClient _client;
     private readonly Log _log;
-    private readonly IMediator _mediator;
     private readonly IReadOnlyList<ITranslationProvider> _translationProviders;
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="MessageCommandExecutedHandler" /> class.
+    /// Initializes a new instance of the <see cref="TranslateMessageCommandHandler" /> class.
     /// </summary>
-    /// <param name="mediator">Mediator to use.</param>
     /// <param name="translationProviders">Translation providers to use.</param>
     /// <param name="client">Discord client to use.</param>
     /// <param name="logger">Logger to use.</param>
-    public MessageCommandExecutedHandler(
-        IMediator mediator,
+    public TranslateMessageCommandHandler(
         IEnumerable<ITranslationProvider> translationProviders,
         IDiscordClient client,
-        ILogger<MessageCommandExecutedHandler> logger
+        ILogger<TranslateMessageCommandHandler> logger
     )
     {
-        _mediator = mediator;
         _translationProviders = translationProviders.ToList();
         _client = client;
         _log = new Log(logger);
@@ -47,15 +37,20 @@ public partial class MessageCommandExecutedHandler
     /// <summary>
     /// Translates the message interacted with to the user's locale.
     /// </summary>
-    /// <param name="request">The request.</param>
+    /// <param name="notification">The notification.</param>
     /// <param name="cancellationToken">The cancellation token.</param>
-    public async Task Handle(ProcessTranslateMessageCommand request, CancellationToken cancellationToken)
+    public async Task Handle(MessageCommandExecutedNotification notification, CancellationToken cancellationToken)
     {
-        if (request.Command.Data.Message.Author.Id == _client.CurrentUser?.Id)
+        if (notification.Command.Data.Name != MessageCommandConstants.TranslateCommandName)
+        {
+            return;
+        }
+
+        if (notification.Command.Data.Message.Author.Id == _client.CurrentUser?.Id)
         {
             _log.TranslatingBotMessageDisallowed();
 
-            await request.Command.RespondAsync(
+            await notification.Command.RespondAsync(
                 "Translating this bot's messages isn't allowed.",
                 ephemeral: true,
                 options: new RequestOptions { CancelToken = cancellationToken }
@@ -64,7 +59,7 @@ public partial class MessageCommandExecutedHandler
             return;
         }
 
-        var sanitizedMessage = FormatUtility.SanitizeText(request.Command.Data.Message.Content);
+        var sanitizedMessage = FormatUtility.SanitizeText(notification.Command.Data.Message.Content);
 
         if (string.IsNullOrWhiteSpace(sanitizedMessage))
         {
@@ -72,7 +67,7 @@ public partial class MessageCommandExecutedHandler
             return;
         }
 
-        var userLocale = request.Command.UserLocale;
+        var userLocale = notification.Command.UserLocale;
 
         string? providerName = null;
         TranslationResult? translationResult = null;
@@ -120,7 +115,7 @@ public partial class MessageCommandExecutedHandler
         if (translationResult == null)
         {
             // Send message if no translation providers support the locale.
-            await request.Command.RespondAsync(
+            await notification.Command.RespondAsync(
                 $"Your locale {userLocale} isn't supported for translation via this action.",
                 ephemeral: true,
                 options: new RequestOptions { CancelToken = cancellationToken }
@@ -133,7 +128,7 @@ public partial class MessageCommandExecutedHandler
         {
             _log.FailureToDetectSourceLanguage();
 
-            await request.Command.RespondAsync(
+            await notification.Command.RespondAsync(
                 "The message couldn't be translated. It might already be in your language or the translator failed to detect its source language.",
                 ephemeral: true,
                 options: new RequestOptions { CancelToken = cancellationToken }
@@ -142,7 +137,7 @@ public partial class MessageCommandExecutedHandler
             return;
         }
 
-        var fromHeading = $"By {MentionUtils.MentionUser(request.Command.Data.Message.Author.Id)}";
+        var fromHeading = $"By {MentionUtils.MentionUser(notification.Command.Data.Message.Author.Id)}";
 
         if (!string.IsNullOrWhiteSpace(translationResult.DetectedLanguageCode))
         {
@@ -160,70 +155,15 @@ public partial class MessageCommandExecutedHandler
 {Format.Bold(toHeading)}:
 {translationResult.TranslatedText}";
 
-        await request.Command.RespondAsync(
+        await notification.Command.RespondAsync(
             embed: new EmbedBuilder()
                 .WithTitle("Translated Message")
-                .WithUrl(GetJumpUrl(request.Command.Data.Message).AbsoluteUri)
+                .WithUrl(GetJumpUrl(notification.Command.Data.Message).AbsoluteUri)
                 .WithDescription(description)
                 .Build(),
             ephemeral: true,
             options: new RequestOptions { CancelToken = cancellationToken }
         );
-    }
-
-    /// <summary>
-    /// Registers the message commands for the guild(s).
-    /// </summary>
-    /// <param name="request">The request.</param>
-    /// <param name="cancellationToken">The cancellation token.</param>
-    public async Task Handle(RegisterMessageCommands request, CancellationToken cancellationToken)
-    {
-        IReadOnlyList<IGuild> guilds =
-            request.Guild != null
-                ? new List<IGuild> { request.Guild }
-                : (
-                    await _client.GetGuildsAsync(options: new RequestOptions { CancelToken = cancellationToken })
-                ).ToList();
-
-        if (!guilds.Any())
-        {
-            return;
-        }
-
-        var translateCommand = new MessageCommandBuilder()
-            .WithName(MessageCommandConstants.TranslateCommandName)
-            .Build();
-
-        foreach (var guild in guilds)
-        {
-            try
-            {
-                await guild.CreateApplicationCommandAsync(
-                    translateCommand,
-                    new RequestOptions { CancelToken = cancellationToken }
-                );
-            }
-            catch (HttpException exception)
-            {
-                _log.FailedToRegisterCommandForGuild(guild.Id, JsonSerializer.Serialize(exception.Errors));
-            }
-        }
-    }
-
-    /// <summary>
-    /// Delegates the MessageCommandExecuted event to the appropriate handler.
-    /// </summary>
-    /// <param name="notification">The notification.</param>
-    /// <param name="cancellationToken">The cancellation token.</param>
-    public async Task Handle(MessageCommandExecutedNotification notification, CancellationToken cancellationToken)
-    {
-        if (notification.Command.Data.Name == MessageCommandConstants.TranslateCommandName)
-        {
-            await _mediator.Send(
-                new ProcessTranslateMessageCommand { Command = notification.Command },
-                cancellationToken
-            );
-        }
     }
 
     /// <summary>
@@ -238,18 +178,12 @@ public partial class MessageCommandExecutedHandler
 
     private sealed partial class Log
     {
-        private readonly ILogger<MessageCommandExecutedHandler> _logger;
+        private readonly ILogger<TranslateMessageCommandHandler> _logger;
 
-        public Log(ILogger<MessageCommandExecutedHandler> logger)
+        public Log(ILogger<TranslateMessageCommandHandler> logger)
         {
             _logger = logger;
         }
-
-        [LoggerMessage(
-            Level = LogLevel.Error,
-            Message = "Failed to register message command for guild ID {guildId} with error(s): {errors}"
-        )]
-        public partial void FailedToRegisterCommandForGuild(ulong guildId, string errors);
 
         [LoggerMessage(Level = LogLevel.Information, Message = "Translating this bot's messages isn't allowed.")]
         public partial void TranslatingBotMessageDisallowed();
